fix: rename to RestaurantsScreen
